All Classes Done And Fixed

+++ prepare/Learning05/Square.cs
@@ -1,13 +1,13 @@
 public class Square : Shape
 {
-    private float _side;
+    private double _side;
 
-    public Square(string color, float side) : base(color)
+    public Square(string color, double side) : base(color)
     {
         _side = side;
     }
 
-    public override float GetArea()
+    public override double GetArea()
     {
         return _side * _side;
     }
